+++ lib/kitchen/figure_element.rb
@@ -53,7 +53,7 @@ module Kitchen
     # @return [Boolean]
     #
     def figure_to_bake?
-      return false if subfigure? || (has_class?('unnumbered') && !has_class?('splash'))
+      return false if subfigure? || (has_class?('unnumbered') && !has_class?('splash') && !caption)
 
       true
     end